Update README.md
It was observed that the Tree sales for both the Fake and Real tree increases from 2010 to 2013 with the Real tree gaining a high rise in sales for the year,but there was a decline in the number of sales in 2014 before further increase till 2016 which was the highest sales so far especially for the Real Tree. Moreover, year 2013 shows a high rise in the number of tree sold for Real Tree followed by year 2011 but the same tread goes for Fake tree as explained earlier. Therefore, the factor responsible for increase in sales and number of tree in 2013 sold should be encouraged for future profit.